More refactoring and javadoc updates
