+++ src/main/java/com/shanebeestudios/skbee/api/event/bound/package-info.java
@@ -1,4 +1,4 @@
 /**
- * {@link org.bukkit.event.Event Events} for {@link com.shanebeestudios.skbee.api.bound.Bound Bounds}
+ * Custom events for {@link com.shanebeestudios.skbee.api.bound.Bound Bounds}
  */
 package com.shanebeestudios.skbee.api.event.bound;
