+++ src/main/java/com/shanebeestudios/skbee/api/event/internal/dialog/PlainMessageEvent.java
@@ -1,4 +1,4 @@
-package com.shanebeestudios.skbee.api.event.dialog;
+package com.shanebeestudios.skbee.api.event.internal.dialog;
 
 import io.papermc.paper.registry.data.dialog.body.PlainMessageDialogBody;
 import org.bukkit.event.Event;
