+++ src/main/java/com/shanebeestudios/skbee/api/event/internal/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Internal events used for structures/sections/etc
+ * <p>
+ * This package is hidden from javadocs
+ * since these events should never be called
+ * </p>
+ */
+package com.shanebeestudios.skbee.api.event.internal;
