+++ src/main/java/com/shanebeestudios/skbee/api/event/internal/worldgen/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Internally used events for ChunkGenerators
+ */
+package com.shanebeestudios.skbee.api.event.internal.worldgen;
