+++ src/main/java/com/shanebeestudios/skbee/api/structure/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Utility and wrapper classes for {@link org.bukkit.structure.Structure Structures}
+ * Utility and wrapper classes for {@link org.bukkit.structure.Structure Structure Templates}
  */
 package com.shanebeestudios.skbee.api.structure;
