--- src/main/java/com/shanebeestudios/skbee/api/worldgen/event/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Internally used events for ChunkGenerators
- */
-package com.shanebeestudios.skbee.api.worldgen.event;
