+++ src/main/java/com/shanebeestudios/skbee/elements/dialog/sections/bodies/SecItemBody.java
@@ -11,8 +11,8 @@ import ch.njol.skript.variables.Variables;
 import ch.njol.util.Kleenean;
 import com.github.shanebeee.skr.Registration;
 import com.shanebeestudios.skbee.api.skript.base.Section;
-import com.shanebeestudios.skbee.api.event.dialog.DialogRegisterEvent;
-import com.shanebeestudios.skbee.api.event.dialog.PlainMessageEvent;
+import com.shanebeestudios.skbee.api.event.internal.dialog.DialogRegisterEvent;
+import com.shanebeestudios.skbee.api.event.internal.dialog.PlainMessageEvent;
 import io.papermc.paper.registry.data.dialog.body.DialogBody;
 import io.papermc.paper.registry.data.dialog.body.ItemDialogBody;
 import io.papermc.paper.registry.data.dialog.body.PlainMessageDialogBody;
