+++ src/main/java/com/shanebeestudios/skbee/elements/dialog/sections/bodies/SecPlainMessageBody.java
@@ -9,8 +9,8 @@ import ch.njol.skript.lang.util.SimpleLiteral;
 import ch.njol.skript.registrations.Classes;
 import ch.njol.util.Kleenean;
 import com.github.shanebeee.skr.Registration;
-import com.shanebeestudios.skbee.api.event.dialog.DialogRegisterEvent;
-import com.shanebeestudios.skbee.api.event.dialog.PlainMessageEvent;
+import com.shanebeestudios.skbee.api.event.internal.dialog.DialogRegisterEvent;
+import com.shanebeestudios.skbee.api.event.internal.dialog.PlainMessageEvent;
 import com.shanebeestudios.skbee.api.skript.base.Section;
 import com.shanebeestudios.skbee.api.wrapper.ComponentWrapper;
 import io.papermc.paper.registry.data.dialog.body.DialogBody;
