+++ src/main/java/com/shanebeestudios/skbee/elements/dialog/sections/inputs/SecSingleOptionInput.java
@@ -9,8 +9,8 @@ import ch.njol.skript.lang.TriggerItem;
 import ch.njol.skript.lang.util.SimpleLiteral;
 import ch.njol.skript.variables.Variables;
 import ch.njol.util.Kleenean;
-import com.shanebeestudios.skbee.api.event.dialog.DialogRegisterEvent;
-import com.shanebeestudios.skbee.api.event.dialog.OptionsEvent;
+import com.shanebeestudios.skbee.api.event.internal.dialog.DialogRegisterEvent;
+import com.shanebeestudios.skbee.api.event.internal.dialog.OptionsEvent;
 import com.github.shanebeee.skr.Registration;
 import com.shanebeestudios.skbee.api.skript.base.Section;
 import com.shanebeestudios.skbee.api.util.Util;
