+++ src/main/java/com/shanebeestudios/skbee/elements/dialog/sections/inputs/SecSingleOptionInputOptions.java
@@ -6,7 +6,7 @@ import ch.njol.skript.lang.Expression;
 import ch.njol.skript.lang.SkriptParser;
 import ch.njol.skript.lang.TriggerItem;
 import ch.njol.util.Kleenean;
-import com.shanebeestudios.skbee.api.event.dialog.OptionsEvent;
+import com.shanebeestudios.skbee.api.event.internal.dialog.OptionsEvent;
 import com.github.shanebeee.skr.Registration;
 import com.shanebeestudios.skbee.api.skript.base.Section;
 import com.shanebeestudios.skbee.api.wrapper.ComponentWrapper;
