+++ src/main/java/com/shanebeestudios/skbee/elements/recipe/effects/EffRecipeSetIngredient.java
@@ -6,8 +6,8 @@ import ch.njol.skript.lang.Expression;
 import ch.njol.skript.lang.SkriptParser.ParseResult;
 import ch.njol.skript.lang.parser.ParserInstance;
 import ch.njol.util.Kleenean;
-import com.shanebeestudios.skbee.api.event.recipe.ShapedRecipeCreateEvent;
-import com.shanebeestudios.skbee.api.event.recipe.ShapelessRecipeCreateEvent;
+import com.shanebeestudios.skbee.api.event.internal.recipe.ShapedRecipeCreateEvent;
+import com.shanebeestudios.skbee.api.event.internal.recipe.ShapelessRecipeCreateEvent;
 import com.shanebeestudios.skbee.api.recipe.RecipeUtil;
 import com.github.shanebeee.skr.Registration;
 import org.bukkit.event.Event;
