+++ src/main/java/com/shanebeestudios/skbee/elements/worldgen/effects/EffChunkDataStructurePlace.java
@@ -4,7 +4,7 @@ import ch.njol.skript.lang.Expression;
 import ch.njol.skript.lang.SkriptParser.ParseResult;
 import ch.njol.util.Kleenean;
 import com.github.shanebeee.skr.Registration;
-import com.shanebeestudios.skbee.api.worldgen.event.BlockPopulateEvent;
+import com.shanebeestudios.skbee.api.event.internal.worldgen.BlockPopulateEvent;
 import com.shanebeestudios.skbee.api.skript.base.Effect;
 import com.shanebeestudios.skbee.api.structure.StructureWrapper;
 import org.bukkit.event.Event;
