+++ src/main/java/com/shanebeestudios/skbee/elements/worldgen/expressions/ExprChunkDataBlock.java
@@ -9,8 +9,8 @@ import ch.njol.util.Kleenean;
 import ch.njol.util.coll.CollectionUtils;
 import com.github.shanebeee.skr.Registration;
 import com.shanebeestudios.skbee.api.skript.base.SimpleExpression;
-import com.shanebeestudios.skbee.api.worldgen.event.BlockPopulateEvent;
-import com.shanebeestudios.skbee.api.worldgen.event.ChunkGenEvent;
+import com.shanebeestudios.skbee.api.event.internal.worldgen.BlockPopulateEvent;
+import com.shanebeestudios.skbee.api.event.internal.worldgen.ChunkGenEvent;
 import org.bukkit.block.data.BlockData;
 import org.bukkit.event.Event;
 import org.bukkit.generator.ChunkGenerator;
