+++ src/main/java/com/shanebeestudios/skbee/elements/worldgen/expressions/ExprChunkDataHighestY.java
@@ -4,8 +4,8 @@ import ch.njol.skript.lang.Expression;
 import ch.njol.skript.lang.SkriptParser.ParseResult;
 import ch.njol.util.Kleenean;
 import com.github.shanebeee.skr.Registration;
-import com.shanebeestudios.skbee.api.worldgen.event.BlockPopulateEvent;
-import com.shanebeestudios.skbee.api.worldgen.event.ChunkGenEvent;
+import com.shanebeestudios.skbee.api.event.internal.worldgen.BlockPopulateEvent;
+import com.shanebeestudios.skbee.api.event.internal.worldgen.ChunkGenEvent;
 import com.shanebeestudios.skbee.api.skript.base.SimpleExpression;
 import org.bukkit.HeightMap;
 import org.bukkit.event.Event;
