+++ src/main/java/com/shanebeestudios/skbee/elements/worldgen/expressions/ExprChunkDataXZ.java
@@ -6,8 +6,8 @@ import ch.njol.skript.lang.SkriptParser.ParseResult;
 import ch.njol.skript.lang.parser.ParserInstance;
 import ch.njol.skript.lang.util.SimpleExpression;
 import ch.njol.util.Kleenean;
-import com.shanebeestudios.skbee.api.worldgen.event.BlockPopulateEvent;
-import com.shanebeestudios.skbee.api.worldgen.event.ChunkGenEvent;
+import com.shanebeestudios.skbee.api.event.internal.worldgen.BlockPopulateEvent;
+import com.shanebeestudios.skbee.api.event.internal.worldgen.ChunkGenEvent;
 import com.github.shanebeee.skr.Registration;
 import org.bukkit.event.Event;
 import org.jetbrains.annotations.NotNull;
